requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
bcrypt==3.2.0
alembic==1.7.1
opencv-python==4.5.3.56
numpy==1.21.2
Pillow==8.3.2
python-multipart==0.0.5
flask-cors==3.0.10
tensorflow==2.6.0
